experimental strip types
